Move back the configuration of hc5 to FeignAutoConfiguration.
