+++ spring-cloud-openfeign-core/src/main/java/org/springframework/cloud/openfeign/FeignAutoConfiguration.java
@@ -26,14 +26,19 @@ import java.util.concurrent.TimeUnit;
 import javax.annotation.PreDestroy;
 
 import com.fasterxml.jackson.databind.Module;
+import feign.AsyncClient;
 import feign.Client;
 import feign.Feign;
 import feign.RequestInterceptor;
+import feign.hc5.ApacheHttp5Client;
+import feign.hc5.AsyncApacheHttp5Client;
 import feign.httpclient.ApacheHttpClient;
 import feign.okhttp.OkHttpClient;
 import okhttp3.ConnectionPool;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.apache.hc.client5.http.impl.async.CloseableHttpAsyncClient;
+import org.apache.hc.client5.http.nio.AsyncClientConnectionManager;
 import org.apache.http.client.HttpClient;
 import org.apache.http.client.config.RequestConfig;
 import org.apache.http.config.RegistryBuilder;
@@ -53,6 +58,8 @@ import org.springframework.cloud.commons.httpclient.ApacheHttpClientConnectionMa
 import org.springframework.cloud.commons.httpclient.ApacheHttpClientFactory;
 import org.springframework.cloud.commons.httpclient.OkHttpClientConnectionPoolFactory;
 import org.springframework.cloud.commons.httpclient.OkHttpClientFactory;
+import org.springframework.cloud.openfeign.clientconfig.AsyncHttpClient5FeignConfigurationHelper;
+import org.springframework.cloud.openfeign.clientconfig.HttpClient5FeignConfigurationHelper;
 import org.springframework.cloud.openfeign.security.OAuth2FeignRequestInterceptor;
 import org.springframework.cloud.openfeign.support.DefaultGzipDecoderConfiguration;
 import org.springframework.cloud.openfeign.support.FeignHttpClientProperties;
@@ -261,6 +268,78 @@ public class FeignAutoConfiguration {
 		@ConditionalOnMissingBean(Client.class)
 		public Client feignClient(okhttp3.OkHttpClient client) {
 			return new OkHttpClient(client);
+		}
+
+	}
+
+	@Configuration(proxyBeanMethods = false)
+	@ConditionalOnClass(ApacheHttp5Client.class)
+	@ConditionalOnMissingBean(org.apache.hc.client5.http.impl.classic.CloseableHttpClient.class)
+	@ConditionalOnProperty("feign.httpclient5.enabled")
+	protected static class HttpClient5FeignConfiguration {
+
+		private org.apache.hc.client5.http.impl.classic.CloseableHttpClient httpClient5;
+
+		@Bean
+		@ConditionalOnMissingBean(org.apache.hc.client5.http.io.HttpClientConnectionManager.class)
+		public org.apache.hc.client5.http.io.HttpClientConnectionManager connectionManager(
+				FeignHttpClientProperties httpClientProperties) {
+			return HttpClient5FeignConfigurationHelper.connectionManager(httpClientProperties);
+		}
+
+		@Bean
+		public org.apache.hc.client5.http.impl.classic.CloseableHttpClient httpClient(
+				org.apache.hc.client5.http.io.HttpClientConnectionManager connectionManager,
+				FeignHttpClientProperties httpClientProperties) {
+			this.httpClient5 = HttpClient5FeignConfigurationHelper.httpClient(connectionManager, httpClientProperties);
+			return this.httpClient5;
+		}
+
+		@Bean
+		@ConditionalOnMissingBean(Client.class)
+		public Client feignClient(org.apache.hc.client5.http.impl.classic.CloseableHttpClient httpClient5) {
+			return new ApacheHttp5Client(httpClient5);
+		}
+
+		@PreDestroy
+		public void destroy() {
+			HttpClient5FeignConfigurationHelper.destroy(this.httpClient5);
+		}
+
+	}
+
+	@Configuration(proxyBeanMethods = false)
+	@ConditionalOnClass(AsyncApacheHttp5Client.class)
+	@ConditionalOnMissingBean(CloseableHttpAsyncClient.class)
+	@ConditionalOnProperty("feign.asynchttpclient5.enabled")
+	protected static class AsyncHttpClient5FeignConfiguration {
+
+		private CloseableHttpAsyncClient asyncHttpClient5;
+
+		@Bean
+		@ConditionalOnMissingBean(AsyncClientConnectionManager.class)
+		public AsyncClientConnectionManager connectionManager(FeignHttpClientProperties httpClientProperties) {
+			return AsyncHttpClient5FeignConfigurationHelper.connectionManager(httpClientProperties);
+		}
+
+		@Bean
+		public CloseableHttpAsyncClient httpClient(AsyncClientConnectionManager connectionManager,
+				FeignHttpClientProperties httpClientProperties) {
+			this.asyncHttpClient5 = AsyncHttpClient5FeignConfigurationHelper.httpClient(connectionManager,
+					httpClientProperties);
+			this.asyncHttpClient5.start();
+			return this.asyncHttpClient5;
+		}
+
+		@Bean
+		@ConditionalOnMissingBean(AsyncClient.class)
+		public AsyncClient feignClient(CloseableHttpAsyncClient httpAsyncClient) {
+			return new AsyncApacheHttp5Client(httpAsyncClient);
+		}
+
+		@PreDestroy
+		public void destroy() {
+			AsyncHttpClient5FeignConfigurationHelper.destroy(this.asyncHttpClient5);
 		}
 
 	}
